+++ src/lib/components/default/default.module.ts
@@ -1,6 +1,7 @@
 import { NgModule }           from '@angular/core';
 import { CommonModule }  from '@angular/common';
 import { DefaultComponent } from './default.component';
+
 export * from './default.component';
 
 @NgModule({